+++ src/test/groovy/org/gradlex/javamodule/dependencies/test/LocalModuleMappingsTest.groovy
@@ -51,6 +51,6 @@ class LocalModuleMappingsTest extends Specification {
         '''
 
         then:
-        build()
+        runner(false, 'build').build()
     }
 }